Refactored all activity classes into an activityClass package. made an abstract class called hamburger activity to implement openHamburgerBar. Adjusted the hamburger bar to implement a person's name
