+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.Design.Light.NoActionBar">
 
         <activity
-            android:name=".LoginActivity"
+            android:name=".ActivityClass.LoginActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -25,25 +25,25 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".StubActivity">
+            android:name=".ActivityClass.StubActivity">
         </activity>
         <activity
-            android:name=".EventActivity">
+            android:name=".ActivityClass.EventActivity">
         </activity>
         <activity
-            android:name=".GroupActivity">
+            android:name=".ActivityClass.GroupActivity">
         </activity>
         <activity
-            android:name=".ProfileActivity">
+            android:name=".ActivityClass.ProfileActivity">
         </activity>
         <activity
-            android:name=".MainActivity">
+            android:name=".ActivityClass.MainActivity">
         </activity>
         <activity
-            android:name=".RegistrationActivity">
+            android:name=".ActivityClass.RegistrationActivity">
         </activity>
-        <activity android:name=".CreateGroupActivity"></activity>
-        <activity android:name=".JoinGroupActivity"></activity>
+        <activity android:name=".ActivityClass.CreateGroupActivity"></activity>
+        <activity android:name=".ActivityClass.JoinGroupActivity"></activity>
     </application>
 
 </manifest>
