+++ app/src/main/java/com/cultivate/juniordesign/cultivate/ActivityClass/GroupActivity.java
@@ -1,4 +1,4 @@
-package com.cultivate.juniordesign.cultivate;
+package com.cultivate.juniordesign.cultivate.ActivityClass;
 
 
 import android.content.Intent;
@@ -13,24 +13,38 @@ import android.widget.PopupWindow;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.google.firebase.auth.FirebaseAuth;
+import com.cultivate.juniordesign.cultivate.Account;
+import com.cultivate.juniordesign.cultivate.R;
 
 /**
  * Created by Paul on 9/10/2017.
  */
 
-public class ProfileActivity extends AppCompatActivity {
+public class GroupActivity extends HamburgerActivity {
+
     private Account user = null;
+    private boolean isAdmin = false;
+    int output = View.VISIBLE;
+    TextView adminTextView;
+
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        user = getIntent().getParcelableExtra("curUser");
-        setContentView(R.layout.profile_stub);
-        FirebaseAuth mAuth = FirebaseAuth.getInstance();
-        String email = mAuth.getCurrentUser().getEmail();
 
-        TextView view = (TextView) findViewById(R.id.textView3);
-        view.setText("Email: " + email);
+        user = getIntent().getParcelableExtra("curUser");
+        setContentView(R.layout.activity_group);
+        isAdmin = checkIfAdmin();
+        //TextView adminTextView  = (TextView)findViewById(R.id.adminTextView);
+        //adminTextView.setVisibility(output);
+    }
+
+    /**
+     * Queries the database to determine if the user is an admin for this group. If so, return true
+     * return boolean whether or not the user is an admin
+     */
+    private boolean checkIfAdmin() {
+        return true;
 
     }
 
@@ -41,21 +55,7 @@ public class ProfileActivity extends AppCompatActivity {
     }
 
     public void openHamburgerBar(View v) {
-        // get a reference to the already created main layout
-        ConstraintLayout mainLayout = (ConstraintLayout) findViewById(R.id.profile_layout);
-
-        // inflate the layout of the popup window
-        LayoutInflater inflater = (LayoutInflater) getSystemService(LAYOUT_INFLATER_SERVICE);
-        View popupView = inflater.inflate(R.layout.hamburgerbar_popup, null);
-
-        // create the popup window
-        int width = LinearLayout.LayoutParams.WRAP_CONTENT;
-        int height = LinearLayout.LayoutParams.WRAP_CONTENT;
-        boolean focusable = true; // lets taps outside the popup also dismiss it
-        final PopupWindow popupWindow = new PopupWindow(popupView, width, height, focusable);
-
-        // show the popup window
-        popupWindow.showAtLocation(mainLayout, Gravity.TOP | Gravity.LEFT, 0, 0);
+        super.openHamburgerBar(v, user);
     }
 
     public void goToStub(View v) {
@@ -82,6 +82,18 @@ public class ProfileActivity extends AppCompatActivity {
         startActivity(group);
     }
 
+    public void goToCreateGroup(View v) {
+        Intent group = new Intent(this, CreateGroupActivity.class);
+        group.putExtra("curUser", user);
+        startActivity(group);
+    }
+
+    public void goToJoinGroup(View v) {
+        Intent group = new Intent(this, JoinGroupActivity.class);
+        group.putExtra("curUser", user);
+        startActivity(group);
+    }
+
     /*
     goes to the login screen; clears user data
      */
@@ -91,6 +103,6 @@ public class ProfileActivity extends AppCompatActivity {
     }
 
     public void notImplemented(View v) {
-        Toast.makeText(ProfileActivity.this, "This feature is not yet implemented", Toast.LENGTH_SHORT).show();
+        Toast.makeText(GroupActivity.this, "This feature is not yet implemented", Toast.LENGTH_SHORT).show();
     }
 }
