+++ app/src/main/java/com/cultivate/juniordesign/cultivate/ActivityClass/JoinGroupActivity.java
@@ -1,4 +1,4 @@
-package com.cultivate.juniordesign.cultivate;
+package com.cultivate.juniordesign.cultivate.ActivityClass;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,20 +9,24 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.PopupWindow;
+import android.widget.TextView;
 import android.widget.Toast;
+
+import com.cultivate.juniordesign.cultivate.Account;
+import com.cultivate.juniordesign.cultivate.R;
 
 /**
  * Created by Paul on 9/10/2017.
  */
 
-public class CreateGroupActivity extends AppCompatActivity {
+public class JoinGroupActivity extends HamburgerActivity {
     private Account user = null;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         user = getIntent().getParcelableExtra("curUser");
-        setContentView(R.layout.create_group_layout);
+        setContentView(R.layout.join_group_layout);
     }
 
     public void goToHome(View v) {
@@ -32,21 +36,7 @@ public class CreateGroupActivity extends AppCompatActivity {
     }
 
     public void openHamburgerBar(View v) {
-        // get a reference to the already created main layout
-        ConstraintLayout mainLayout = (ConstraintLayout) findViewById(R.id.create_group_layout);
-
-        // inflate the layout of the popup window
-        LayoutInflater inflater = (LayoutInflater) getSystemService(LAYOUT_INFLATER_SERVICE);
-        View popupView = inflater.inflate(R.layout.hamburgerbar_popup, null);
-
-        // create the popup window
-        int width = LinearLayout.LayoutParams.WRAP_CONTENT;
-        int height = LinearLayout.LayoutParams.WRAP_CONTENT;
-        boolean focusable = true; // lets taps outside the popup also dismiss it
-        final PopupWindow popupWindow = new PopupWindow(popupView, width, height, focusable);
-
-        // show the popup window
-        popupWindow.showAtLocation(mainLayout, Gravity.TOP | Gravity.LEFT, 0, 0);
+        super.openHamburgerBar(v, user);
     }
 
     public void goToStub(View v) {
@@ -74,7 +64,7 @@ public class CreateGroupActivity extends AppCompatActivity {
     }
 
     /*
-    goes to the login screen; clears user data  
+    goes to the login screen; clears user data
      */
     public void goToLogout(View v) {
         Intent i = new Intent(this, LoginActivity.class);
@@ -82,6 +72,6 @@ public class CreateGroupActivity extends AppCompatActivity {
     }
 
     public void notImplemented(View v) {
-        Toast.makeText(CreateGroupActivity.this, "This feature is not yet implemented", Toast.LENGTH_SHORT).show();
+        Toast.makeText(JoinGroupActivity.this, "This feature is not yet implemented", Toast.LENGTH_SHORT).show();
     }
 }
