+++ app/src/main/java/com/cultivate/juniordesign/cultivate/ActivityClass/ProfileActivity.java
@@ -1,4 +1,5 @@
-package com.cultivate.juniordesign.cultivate;
+package com.cultivate.juniordesign.cultivate.ActivityClass;
+
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,20 +10,30 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.PopupWindow;
+import android.widget.TextView;
 import android.widget.Toast;
+
+import com.cultivate.juniordesign.cultivate.Account;
+import com.cultivate.juniordesign.cultivate.R;
+import com.google.firebase.auth.FirebaseAuth;
 
 /**
  * Created by Paul on 9/10/2017.
  */
 
-public class JoinGroupActivity extends AppCompatActivity {
+public class ProfileActivity extends HamburgerActivity {
     private Account user = null;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         user = getIntent().getParcelableExtra("curUser");
-        setContentView(R.layout.join_group_layout);
+        setContentView(R.layout.profile_stub);
+        FirebaseAuth mAuth = FirebaseAuth.getInstance();
+        String email = mAuth.getCurrentUser().getEmail();
+
+        TextView view = (TextView) findViewById(R.id.textView3);
+        view.setText("Email: " + email);
+
     }
 
     public void goToHome(View v) {
@@ -32,21 +43,7 @@ public class JoinGroupActivity extends AppCompatActivity {
     }
 
     public void openHamburgerBar(View v) {
-        // get a reference to the already created main layout
-        ConstraintLayout mainLayout = (ConstraintLayout) findViewById(R.id.join_group_layout);
-
-        // inflate the layout of the popup window
-        LayoutInflater inflater = (LayoutInflater) getSystemService(LAYOUT_INFLATER_SERVICE);
-        View popupView = inflater.inflate(R.layout.hamburgerbar_popup, null);
-
-        // create the popup window
-        int width = LinearLayout.LayoutParams.WRAP_CONTENT;
-        int height = LinearLayout.LayoutParams.WRAP_CONTENT;
-        boolean focusable = true; // lets taps outside the popup also dismiss it
-        final PopupWindow popupWindow = new PopupWindow(popupView, width, height, focusable);
-
-        // show the popup window
-        popupWindow.showAtLocation(mainLayout, Gravity.TOP | Gravity.LEFT, 0, 0);
+        super.openHamburgerBar(v, user);
     }
 
     public void goToStub(View v) {
@@ -82,6 +79,6 @@ public class JoinGroupActivity extends AppCompatActivity {
     }
 
     public void notImplemented(View v) {
-        Toast.makeText(JoinGroupActivity.this, "This feature is not yet implemented", Toast.LENGTH_SHORT).show();
+        Toast.makeText(ProfileActivity.this, "This feature is not yet implemented", Toast.LENGTH_SHORT).show();
     }
 }
