+++ app/src/main/res/layout/activity_login.xml
@@ -10,7 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:id="@+id/activity_login"
-    tools:context="com.cultivate.juniordesign.cultivate.LoginActivity">
+    tools:context="com.cultivate.juniordesign.cultivate.ActivityClass.LoginActivity">
 
 
     <TextView
