+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cultivate.juniordesign.cultivate.MainActivity">
+    tools:context="com.cultivate.juniordesign.cultivate.ActivityClass.MainActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/home_toolbar"
