+++ app/src/main/res/layout/activity_registration.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.cultivate.juniordesign.cultivate.RegistrationActivity">
+    tools:context="com.cultivate.juniordesign.cultivate.ActivityClass.RegistrationActivity">
 
 
     <TextView
